requirements.txt
pandas~=1.5.3
openpyxl~=3.1.2
requests~=2.28.2
python-dotenv~=1.0.0
